scraper/requirements.txt
pymongo
python_dotenv
cerberus
scikit-learn
pandas